Remove some casts

+++ src/gtest/exec.rs
@@ -54,7 +54,7 @@ pub fn get_tests<P: Into<PathBuf>>(
     Ok(tests)
 }
 
-pub fn cmd<P: Into<PathBuf>>(test_executable: P, job_index: u64, jobs: u64) -> Command {
+pub fn cmd<P: Into<PathBuf>>(test_executable: P, job_index: usize, jobs: usize) -> Command {
     let mut child = Command::new(&test_executable.into());
 
     child.env("GTEST_SHARD_INDEX", job_index.to_string());
@@ -67,7 +67,7 @@ pub fn cmd<P: Into<PathBuf>>(test_executable: P, job_index: u64, jobs: u64) -> C
 }
 
 pub fn process_shard(
-    shard: u64,
+    shard: usize,
     child: Child,
     sender: Sender<TestResult>,
     done: Sender<()>,
